Added feature flag to app config, added set up method and general tidy up
