+++ config/config.go
@@ -23,6 +23,7 @@ type Configuration struct {
 	HealthCheckInterval      time.Duration `envconfig:"HEALTHCHECK_INTERVAL"`
 	EnablePrivateEnpoints    bool          `envconfig:"ENABLE_PRIVATE_ENDPOINTS"`
 	EnableDetachDataset      bool          `envconfig:"ENABLE_DETACH_DATASET"`
+	EnableCMDAuth            bool          `envconfig:"ENABLE_CMD_AUTH"`
 	MongoConfig              MongoConfig
 }
 
@@ -56,6 +57,7 @@ func Get() (*Configuration, error) {
 		HealthCheckInterval:      30 * time.Second,
 		EnablePrivateEnpoints:    false,
 		EnableDetachDataset:      false,
+		EnableCMDAuth:            false,
 		MongoConfig: MongoConfig{
 			BindAddr:   "localhost:27017",
 			Collection: "datasets",
